+++ fastly/ngwaf/v1/lists/api_get.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 
 	"github.com/fastly/go-fastly/v10/fastly"
+	"github.com/fastly/go-fastly/v10/fastly/ngwaf/v1/common"
 )
 
 // GetInput specifies the information needed for the Get() function to
@@ -13,20 +14,24 @@ import (
 type GetInput struct {
 	// ListID is the workspace identifier (required).
 	ListID *string
-	// WorkspaceID is the workspace identifier (required).
-	WorkspaceID *string
+	// Scope defines where the list is applied, including its type (e.g.,
+	// "workspace" or "account") and the specific IDs it applies to (required).
+	Scope *common.Scope
 }
 
 // Get retrieves the specified list.
 func Get(ctx context.Context, c *fastly.Client, i *GetInput) (*List, error) {
-	if i.WorkspaceID == nil {
-		return nil, fastly.ErrMissingWorkspaceID
-	}
 	if i.ListID == nil {
 		return nil, fastly.ErrMissingListID
 	}
+	if i.Scope == nil {
+		return nil, fastly.ErrMissingScope
+	}
 
-	path := fastly.ToSafeURL("ngwaf", "v1", "workspaces", *i.WorkspaceID, "lists", *i.ListID)
+	path, err := common.BuildPath(i.Scope, "lists", *i.ListID)
+	if err != nil {
+		return nil, fmt.Errorf("failed to build API path: %w", err)
+	}
 
 	resp, err := c.Get(ctx, path, fastly.CreateRequestOptions())
 	if err != nil {